Updated project architecture, just moved some code where it belongs in the MVC pattern
