+++ views/cmd/remove.go
